backend/requirements.txt
fastapi[all]==0.128.1
SQLAlchemy==2.0.48
PyMySQL==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
uvicorn==0.40.0
python-dotenv==1.1.1